Small reorganization of preferences and sessions status resolving.
